JIF-98 - Refactored profile screen to use new Gender model.

+++ Jifflr-iOS/Helpers/Dashboard/DashboardManager.swift
@@ -29,7 +29,7 @@ class DashboardManager: NSObject {
                     dashboardStats.adsViewed = adsViewed
                 }
 
-                if let money = dashboardJSON["money"] as? Int {
+                if let money = dashboardJSON["money"] as? Double {
                     dashboardStats.money = money
                 }
